# Preallocate sets before filling them with endpoints